+++ pol-core/bscript/eprog3.cpp
@@ -64,7 +64,7 @@ void EScriptProgram::dump_casejmp( ostream& os, const Token& token )
         dataptr += 1;
         if (type == CASE_TYPE_LONG)
         {
-            unsigned long lval = * (const unsigned long*) dataptr;
+            unsigned int lval = * (const unsigned int*) dataptr;
             dataptr += 4;
             os << "\t" << lval << ": @" << offset << endl;
         }
